+++ amazon-order-scraper/__init__.py
@@ -1,1 +1,1 @@
-from .scrape import parse_amazon_transactions_for_year
+from .controller import parse_amazon_transactions_for_year
